refactor of prover and verifier functions
